Rename util.rs to sql.rs
