+++ src/adapter/src/coord.rs
@@ -115,7 +115,6 @@ use crate::coord::id_bundle::CollectionIdBundle;
 use crate::coord::peek::PendingPeek;
 use crate::coord::read_policy::{ReadCapability, ReadHolds};
 use crate::coord::timeline::{TimelineState, WriteTimestamp};
-pub(crate) use crate::coord::util::CoordTimestamp;
 use crate::error::AdapterError;
 use crate::session::Session;
 use crate::sink_connection;
@@ -133,9 +132,9 @@ mod indexes;
 mod message_handler;
 mod read_policy;
 mod sequencer;
+mod sql;
 mod timeline;
 mod timestamp_selection;
-mod util;
 
 /// The default is set to a second to track the default timestamp frequency for sources.
 pub const DEFAULT_LOGICAL_COMPACTION_WINDOW_MS: Option<u64> = Some(1_000);
@@ -868,5 +867,51 @@ pub async fn serve<S: Append + 'static>(
             Ok((handle, client))
         }
         Err(e) => Err(e),
+    }
+}
+
+pub trait CoordTimestamp:
+    timely::progress::Timestamp
+    + timely::order::TotalOrder
+    + differential_dataflow::lattice::Lattice
+    + std::fmt::Debug
+{
+    /// Advance a timestamp by the least amount possible such that
+    /// `ts.less_than(ts.step_forward())` is true. Panic if unable to do so.
+    fn step_forward(&self) -> Self;
+
+    /// Advance a timestamp forward by the given `amount`. Panic if unable to do so.
+    fn step_forward_by(&self, amount: &Self) -> Self;
+
+    /// Retreat a timestamp by the least amount possible such that
+    /// `ts.step_back().unwrap().less_than(ts)` is true. Return `None` if unable,
+    /// which must only happen if the timestamp is `Timestamp::minimum()`.
+    fn step_back(&self) -> Option<Self>;
+
+    /// Return the maximum value for this timestamp.
+    fn maximum() -> Self;
+}
+
+impl CoordTimestamp for mz_repr::Timestamp {
+    fn step_forward(&self) -> Self {
+        match self.checked_add(1) {
+            Some(ts) => ts,
+            None => panic!("could not step forward"),
+        }
+    }
+
+    fn step_forward_by(&self, amount: &Self) -> Self {
+        match self.checked_add(*amount) {
+            Some(ts) => ts,
+            None => panic!("could not step {self} forward by {amount}"),
+        }
+    }
+
+    fn step_back(&self) -> Option<Self> {
+        self.checked_sub(1)
+    }
+
+    fn maximum() -> Self {
+        Self::MAX
     }
 }
