+++ src/adapter/src/coord/sequencer.rs
@@ -2919,4 +2919,13 @@ impl<S: Append + 'static> Coordinator<S> {
         let revision = ps.catalog_revision;
         session.create_new_portal(sql, desc, plan.params, Vec::new(), revision)
     }
+
+    pub(crate) fn allocate_transient_id(&mut self) -> Result<GlobalId, AdapterError> {
+        let id = self.transient_id_counter;
+        if id == u64::MAX {
+            coord_bail!("id counter overflows i64");
+        }
+        self.transient_id_counter += 1;
+        Ok(GlobalId::Transient(id))
+    }
 }
